refactor(session): extract breakpoint tooling into src/session/breakpoints/
Introduces `OperationsContext` — the late-bound view of the operations facade
that every collaborator gets — and moves the first slice behind it:

- `breakpoints/breakpoint-controller.ts` owns set/remove/clear plus the two
  replace-all DAP re-sends and the live-sync warning.
- `breakpoints/anchor-resolution.ts` holds `reresolveAnchors` as a free
  function; restart and the JVM hot swap both call it.
- `breakpoints/launch-warnings.ts` holds the three warning builders as pure
  functions. Their tests now import them instead of invoking them off
  `SessionManager.prototype` with a bare receiver.

Context members are arrows and getters, never captured references, so tests
that reassign `selectPolicy` or write the tunables on a live instance still
reach the collaborators. `selectPolicy` (data layer) and `selectStorePolicy`
(session store, throws on unknown) stay distinct; the function-breakpoint
launch warning keeps reading the store's, behind its try/catch.
